Updates to show songs for a user
Uppdates in API Repo (Song should include artist and genre), song-VM (artist and genre are there, delete Icolletion for artist)
Updated MapPost endpoint in API so it takes artistId and genreId, not beautiful but works.
Updated APIModel for ListSongs in Console client, it includes artist and genre now.
Implementet GetSongsForUser and GetGenresForUser in console client, draft for ConnectSongToUser but too many questions so need to wait for our meeting